2/11/2025

+++ Neural Pipeline/get_matching_intan_blackrock.m
@@ -4,11 +4,18 @@ clc
 %% read ccf files
 % Get the current directory
 currentDirectory = pwd;
-ccfdir  = uigetdir('\\10.16.59.34\cullenlab_server\Current Project Databases - NHP\2021 Abducens Stimulation (Neuropixel)\Data\Project 1 - Occulomotor Kinematics\', 'choose root folder for ccf files');
+ccfdir  = uigetdir('\\10.16.59.34\cullenlab_server\Current Project Databases - NHP\2021 Abducens Stimulation (Neuropixel)\Data\Project 1 - Occulomotor Kinematics\', 'choose root folder for blackrock files');
 intandir  = uigetdir('\\10.16.59.34\cullenlab_server\Current Project Databases - NHP\2021 Abducens Stimulation (Neuropixel)\Data\Project 1 - Occulomotor Kinematics\', 'choose root folder for intan files');
-datafolder = uigetdir('C:\','choose an output data folder');
+datafolder = uigetdir(pwd,'choose an output data folder');
 % Define the output Excel file path
-outputFile = fullfile(currentDirectory, 'CCF_Files.xlsx');
+%%
+prompt = {'Copy the sorted .rhs files to a separate folder (Y/N) : '};
+dlgtitle = 'Input';
+fieldsize = [1 45];
+definput = {'Y'};
+flag = inputdlg(prompt,dlgtitle,fieldsize,definput);
+%%
+outputFile = fullfile(datafolder, 'CCF_Files.xlsx');
 
 % Search for '.ccf' files in the current directory and subfolders
 files = dir(fullfile(ccfdir, '**', '*.ccf'));
@@ -56,24 +63,32 @@ end
 T = cell2table(data, 'VariableNames', {'SubfolderName', 'FileName', 'FullPath', 'LastModified'});
 
 % Write to Excel file
-outputFile = fullfile(currentDirectory, 'Output.xlsx');
+outputFile = fullfile(datafolder, 'Output.xlsx');
 writetable(T, outputFile, 'Sheet', 'Setting Files');
 
 disp('Export complete. The file is saved as Output.xlsx in the current directory.');
 
 %%
+cd(datafolder)
 CCFFiles = readtable('CCF_Files.xlsx');
 Output = readtable('Output.xlsx');
-datesA = datetime(CCFFiles.LastModified);  % Including the time
-datesB = datetime(Output.LastModified);  % Including the time
-timeA = dateshift(datesA, 'start', 'minute');  % Remove day, keep time part
-timeB = dateshift(datesB, 'start', 'minute');  % Same for datesB
-timeA = datetime(timeA, 'Format', 'HH:mm');  % Only keep hours and minutes
-timeB = datetime(timeB, 'Format', 'HH:mm');  % Only keep hours and minutes
-[isMatch, idxB] = ismember(datesA, datesB);  % Check if times in A match with B
 %%
-matchesA = timeA(isMatch);
-matchesB = timeB(idxB(idxB~=0));
+datesA = datetime(CCFFiles.LastModified, 'Format','HH:mm:ss');  % Including the time
+datesB = datetime(Output.LastModified, 'Format','HH:mm:ss');  % Including the time
+% Convert datetime to numeric (in minutes since a reference time)
+refTime = min([datesA; datesB]);  % Use the earliest datetime as reference
+numA = minutes(datesA - refTime);
+numB = minutes(datesB - refTime);
+
+% Define tolerance 
+temp = max(abs([numA(:);numB(:)]));
+
+tolerance =0.5/temp;
+
+[isMatch, idxB] = ismembertol(numA, numB, tolerance);  % Check if times in A match with B
+%%
+matchesA = datesA(isMatch);
+matchesB = datesB(idxB(isMatch));
 
 %%
 filesA = CCFFiles.FileName(isMatch);
@@ -84,11 +99,21 @@ filename = 'matches.xlsx';
 writetable(T,filename,'Sheet','matched')
 %%
 
-cd(datafolder)
-for i =  34: length(filesB)
-    file = filesB(i);
-    folder_path = fullfile('\\10.16.59.34\cullenlab_server\Current Project Databases - NHP\2021 Abducens Stimulation (Neuropixel)\Data\Project 1 - Occulomotor Kinematics\Daphne_Session_3 - Copy\Intan', file);
-    [~,file_name,~] = fileparts(filesA(i));
-    mkdir(file_name);
-    copyfile(folder_path, file_name)
+%%
+outputfolder = fullfile(datafolder, 'Intan_Sorted');
+if ~exist( outputfolder, 'dir')
+    mkdir(outputfolder)
+end
+%%
+cd(outputfolder)
+for i =  1: length(filesB)
+    file = filesB{i};
+    folder_path = fullfile(intandir, file, '*.rhs');
+    [~,file_name,~] = fileparts(filesA{i});
+    disp(['copying .rhs from BR folder ' file_name '...'])
+    if ~exist(file_name, 'file')
+        mkdir(file_name);
+        copyfile(folder_path, file_name)
+    end
+    
 end
